Move live-variant fixtures into conftest.
Avoid importing the tests package so pyrefly can type-check under the
src layout.
