+++ tests/test_async_client.py
@@ -1,6 +1,7 @@
 """Tests for the async CoderPad client."""
 
 import json
+from collections.abc import Callable
 from http import HTTPStatus
 from pathlib import Path
 from urllib.parse import parse_qs
@@ -21,7 +22,6 @@ from coderpad.types import (
     QuestionFileContent,
     SortOrder,
 )
-from tests import LIVE_VARIANT_ORGANIZATION_ID, live_variant_response
 
 
 class TestAsyncCoderPad:
@@ -338,7 +338,10 @@ class TestAsyncGetPadEnvironment:
 
     @staticmethod
     @pytest.mark.asyncio
-    async def test_live_response_variants() -> None:
+    async def test_live_response_variants(
+        live_variant_organization_id: int,
+        live_variant_response: Callable[..., TransportResponse],
+    ) -> None:
         """Undocumented environment and organization variants are
         supported.
         """
@@ -364,7 +367,7 @@ class TestAsyncGetPadEnvironment:
         assert environment.file_contents[0].binary is True
 
         organization = await client.organization.get()
-        assert organization.id == LIVE_VARIANT_ORGANIZATION_ID
+        assert organization.id == live_variant_organization_id
         assert organization.single_sign_in_url is None
